bot/requirements.txt
playwright==1.50.0
langchain-google-genai
browser-use
python-dotenv
langchain
google-genai
tenacity
